extract web3.storage hook
